Dev: Decided services should not contain business logic, as that is usually for mvc applications.
