+++ Benny-Scraper/Benny-Scraper.csproj
@@ -9,7 +9,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Aspose.HTML" Version="23.2.0" />
+    <PackageReference Include="Epub.Net" Version="1.0.16-beta" />
+    <PackageReference Include="EPubFactory" Version="1.1.0" />
     <PackageReference Include="HtmlAgilityPack" Version="1.11.46" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="7.0.1">
       <PrivateAssets>all</PrivateAssets>
