+++ Benny-Scraper/NovelPageScraper.cs
@@ -1,6 +1,4 @@
-﻿using Aspose.Html;
-using Aspose.Html.Saving;
-using Benny_Scraper.Interfaces;
+﻿using Benny_Scraper.Interfaces;
 using Benny_Scraper.Models;
 using HtmlAgilityPack;
 using System.Net;
