+++ Benny-Scraper/Program.cs
@@ -8,6 +8,8 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using OpenQA.Selenium;
+using Benny_Scraper.BusinessLogic;
+using Benny_Scraper.BusinessLogic.Interfaces;
 
 //[assembly: log4net.Config.XmlConfigurator(ConfigFile = "log4net.config", Watch = true)]
 
@@ -31,11 +33,12 @@ namespace Benny_Scraper
             dbInitializer.Initialize();
 
             INovelService novelService = host.Services.GetRequiredService<INovelService>();
+            INovelProcessor novelProcessor = host.Services.GetRequiredService<INovelProcessor>();
 
             // Uri help https://www.dotnetperls.com/uri#:~:text=URI%20stands%20for%20Universal%20Resource,strings%20starting%20with%20%22http.%22
             Uri novelTableOfContentUri = new Uri("https://novelfull.com/paragon-of-sin.html");
 
-            await novelService.ProcessNovelAsync(novelTableOfContentUri);
+            await novelProcessor.ProcessNovelAsync(novelTableOfContentUri);
         }
     }
 }
